Some much needed refactoring
