+++ lib/Semaphore.js
@@ -1,0 +1,29 @@
+function Semaphore(initialCount) {
+	this._count = initialCount || 1;
+	this._waiting = [];
+}
+
+module.exports = Semaphore;
+
+
+Semaphore.prototype.signal = function () {
+	this._count += 1;
+
+	if (this._count <= 0) {
+		var waiter = this._waiting.shift();
+		if (waiter) {
+			waiter();
+		}
+	}
+};
+
+
+Semaphore.prototype.wait = function (cb) {
+	this._count -= 1;
+
+	if (this._count < 0) {
+		this._waiting.push(cb);
+	} else {
+		cb();
+	}
+};
